Modify view - Item.hpp - GrowthItem.hpp - PoisonItem.hpp - GameManager.hpp - rules.txt
